use precompile headers for win32
git-svn-id: file:///home/unikey/trunk@46 e1ceb1ec-800a-4d3f-98f3-0cab0dcde3af

+++ src/ukengine/mactab.cpp
@@ -20,7 +20,7 @@
  * Free Software Foundation, Inc., 59 Temple Place - Suite 330,
  * Boston, MA 02111-1307, USA.
  */
-
+#include "stdafx.h"
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
@@ -108,7 +108,7 @@ int CMacroTable::loadFromFile(const TCHAR *fname)
   if (f == NULL) 
     return 0;
   char line[MAX_MACRO_LINE];
-  int len;
+  size_t len;
 
   resetContent();
   while (fgets(line, sizeof(line), f)) {
